requirements.txt
Flask==2.2.3
python-docx==0.8.11
firebase-admin==6.0.1
gunicorn==20.1.0
